Refactored Repository
